pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "ipykernel>=7.1.0",
    "langchain[google-genai]>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "psycopg2>=2.9.11",
    "streamlit>=1.52.2",
]
